Move micado directory to top level
This should help with making SourceDescriptionFitsKeywords from ScopeSim
work, because the full package paths are now shorter.

Not all micado sources are (always) calibration sources, so they
should not be in the calibration directory anyway.
