+++ scopesim_templates/calibration/__init__.py
@@ -1,2 +1,1 @@
 from .calibration import *
-from . import micado
