+++ scopesim_templates/micado/cluster.py
