+++ scopesim_templates/micado/flatlamp.py
