+++ scopesim_templates/micado/viking_fields/__init__.py
