+++ scopesim_templates/micado/viking_fields/cat_science.dat
