+++ scopesim_templates/micado/viking_fields/cat_stdstar.dat
