+++ scopesim_templates/micado/viking_fields/faintgal_1.dat
