+++ scopesim_templates/micado/viking_fields/faintgal_2.dat
