+++ scopesim_templates/micado/viking_fields/faintgal_3.dat
