+++ scopesim_templates/micado/viking_fields/field_illum.eps
