+++ scopesim_templates/micado/viking_fields/field_stdstar.eps
